Add support for step matching hook
This patch was ported from our old CukeBins fork. It won't ever be
merged into the upstream master, but it must be in our fork's master
until we are able to implement a better solution for this problem in
our other repositories. See also:

https://github.com/cucumber/cucumber-cpp/pull/149

Once an appropriate workaround is found, this patch should be
reverted.